Use constructor to set properties
Use the constructor to let the object be properly initialized with valid
data when it is created.  The constructor centralizes the initialization
logic, making it easier to manage and reuse.  Without the constructor,
we would need to manually set the properties after creating the object.

Pass the object as independent bits in {{ }} to mix in with Blade view
in lieu of PHP block.