Add addTax and applyDiscount to modify amount
Build addTax and applyDiscount methods in Transaction class, and call
them in the controller index method to modify the amount property.

The self return type indicates that the method returns an instance of
the same class (Transaction).  This enhances method chaining.  The order
of chaining doesn't affect the final amount.

Use number_format PHP function for two decimal places of the amount.