Use nullsafe operator
The nullsafe operator (?->) was introduced in PHP 8.0 to make it easier
and cleaner to access properties or call methods on objects that may be
null.

It works like the regular object operator (->), but if the left-hand
side is null, it short-circuits and returns null instead of throwing an
error.

Without the nullsafe operator, accessing deeply nested properties can be
risky and it will throw an error if any part of the chain is null (e.g.,
$item->customer is null).

In our example, if $item->customer or $item->customer->paymentProfile is
null, the whole expression evaluates to null.