Add status constants in Transaction class
Build a list of status constants in Transaction class using the const
keyword.  Declare the constants as public since we intend to reference
them from outside the class.

Use fully qualified class name and the scope resolution operator (::) to
access the class constants in the Blade view.  This shows why we declare
the constants public.