Use ::class for full namespace of a class
Using ::class is a clean and reliable way to get the full namespace of a
class.  It's better than hard coded string of 'App\Accounts\Transaction'
for auto-resolving namespaces, refactoring safely and preventing typos.

This Is perfectly valid and serves as a great debugging or documentation
tool in the view.  In the rendered page we deliver this message:

"We're dealing with the App\Accounts\Transaction class here."

That’s super helpful during development.