Add a new status
Use setStatus method to override the default value of status property
and rebuild the view to display the new status.

The use of Status::PENDING->value and Status::DECLINED->value in lieu of
hard coded strings of 'pending' and 'declined' avoids magic strings that
can lead to bugs if we mistype them ('pendign', 'pended', etc.).  This
also provides centralized definition.  If later we want to change, for
instance, 'pending' to 'in_progress', we only need to update the string
in one place, not 'pending' all over the places.