Add instance property to hold dynamic value of static id
Use the static $id property to track the count of Transaction instances.
For each count use it as the value of the new $transactionId property of
the initialized instance.  The value of $id will change as the number of
instances change, but the $transactionId is locked by the value of count
when the instance is initialized.