Build toastWithButter method in ToasterDeluxe
Replicate the toast method with butter added and form toastWithButter in
the ToasterDeluxe class.

Use both toast and toastWithButter methods in the Toaster controller.
Substitute the compact helper with the manual array to overcome the
complexity.  The compact is only useful for passing simple data, but not
for complicated one.

Show the buttered slices in the same way as the regular slices and apply
snake_case in alignment to Laravel convention for collaboration.  We can
now use $butter_slices rather than $butterSlices by benefiting from the
manual array.