Use Laravel dependency injection
Dependency Injection (DI) means that instead of hard coding dependencies
inside a class or method, they are passed in from the outside.

This makes the code flexible and loosely coupled.  We can change
behavior without changing code and classes don’t rely on concrete
implementations.

When we type-hint classes in a controller method, Laravel uses its
Service Container to automatically resolve and inject the dependencies.