Use Laravel service container binding
Laravel uses something called bindings in its service container to know
which concrete class should be used when an interface is type-hinted.

The use of bind method serves this:

When we ask for DebtCollector, give us an instance of FastAgency.

If later we decide to use a different one, we only need to update the
service container for the binding.  No controller code changes needed
and everything still works!

In PHP, constants are resolved at compile time, based on the class we
reference, but not the actual object instance.

Even though DebtCollector is bound to FastAgency, if we access by the
constant via DebtCollector::FEE, PHP looks at the interface's constant,
but not the implementing class.  And this is by design.  Constants are
not polymorphic, and they’re statically bound.

To get around this, use instance name rather than the class name.  But
is it the best solution?