Use an additional constant of EXPEDITED_FEE
Let a constant be constant.  It's a terrible idea to change the constant
and that defeats its purpose.  If a constant needs changing, it's much
better to use a new one.

Build a new method to put the constants to use.  It's much better and
safe to wrap constants inside a method.  While the use of $agency::FEE
is brilliant, delegating it to the method is mastery.

It is important by not mixing access types through self::EXPEDITED_FEE
and self::FEE.  It's consistent and scalable by using self::FEE, which,
in our current example, is much better than DebtCollector::FEE.  Outside
our example, it might be different from DebtCollector::FEE and it will
be adjusted accordingly.