Build __isset magic method
Build the __isset magic method that we can use isset($invoice->amount)
and can check whether the key exists in $this->attributes.  The magic
method is perfect for dynamic property access.  This means that this
code in Blade works perfectly:

  {{ isset($invoice->amount) ? 'Unpaid' : 'Paid' }}

In the controller this line: $invoice->amount; does nothing.  It reads
the value, but doesn't assign anything.  So if we're trying to simulate
an unpaid invoice (amount is set), or paid one (amount not set), this
won’t help.  Removing it is the best way to go.

Add the span element and use Tailwind to make the isset helper look
great and in alignment with the rest of the view.