Build the amount dynamic property of Invoice
Build the dynamic property called amount in the $invoice object.  But in
the class definition, there's no public $amount; or any other declared
property.  So what happens?

PHP calls the __set() magic method as we defined.  Behind the scenes,
__set() magic method turns $invoice->amount = 5000; into:

  $invoice->__set('amount', 5000);

Which stores the value like this inside the Invoice object:

  $this->attributes['amount'] = 5000;

Even though amount is not a real property of the class, it gets stored
dynamically in $this->attributes.  And the Blade view dutifully displays
it as unpaid.