Build __unset magic method
Build __unset method so that we can use unset method to remove dynamic
property.  When we run unset($invoice->amount), behind the scenes, this
calls __unset magic method as we defined.

With the addition of unset call, now isset($invoice->amount) becomes
false and the Blade view will show "Paid".