Build static method of __callStatic
Build static helper method of mapStaticCall and use it to build static
method of __callStatic.

Use self keyword for the static class where the method is defined.  Yet
the method can't be overridden in the subclass.

Pass the result to the view and display it in the view in alignment with
the rest of the view.