Use static keyword for the class
When we use 'self', we refer to the class where the method is defined.
The late static binding is not applied.  On the other hand, when we use
'static', we refer to the class that was called at runtime, for example,
in the context of a subclass.  And that is what late static binding is.

So for the __callStatic() method, if we plan to extend the Invoice class
later (e.g., PaidInvoice, UnpaidInvoice), then we use this:

  static::mapStaticCall($name, $arguments)

Because we want the child class’s version if it exists.  That follows
the best practice for the static magic method of __callStatic().