Use methods for self and static scope resolution
In a parent-child class relationship, from Invoice::getClassBySelf(),
self always refers to the class where the method is defined, the Bill
class.  But from Invoice::getClassByStatic(), static respects late
static binding and uses the called class, the Invoice class.

Build the view and pass the return value of the above methods to it for
the demo of late static binding with static vs. compile-time resolution
with self.