Fix hard coded string
Instead of using hard coded strings of 'Bill' and 'Invoice', we can use
Reflection approach, which is the cleanest and most maintainable.  It is
slightly slower, but for most applications the difference is negligible.

The __toString() magic method can't be declared with static.  This also
makes it very attractive to use $this keyword — anther reason to go with
Reflection approach.