Build static methods of early and late binding
Use self::class and static::class for methods of early binding and late
binding.  Pass the call of the methods with Invoice:: prefix to the view
for the effects of early and late bindings.

These methods work the same way as getClassName() and its helper, but we
don't have to use __CLASS__ constant.  The helper is spared as well.