Use type hinting for show() method
Type hinting is rendered through the argument of NamedObject $billing,
which expects this argument to be an instance of NamedObject or one of
its subclasses of this hierarchy:

  abstract class NamedObject { ... }
  class Bill extends NamedObject { ... }
  class Invoice extends Bill { ... }

With the above hierarchy, it interprets Invoice like this:

1) Invoice is a Bill
2) Bill is a NamedObject
3) Therefore, Invoice is a NamedObject

When Laravel injects either new Invoice() or new Bill(), it satisfies
the type hint.

When we do: GET /billings/process

Laravel will:

1) Match {billing} as 'process'
2) Use the resolver to return new Invoice()
3) Pass that object to the controller like this:

  public function show(NamedObject $billing)
  // $billing = new Invoice()

This works because Invoice is a subclass of NamedObject.

The role of getViewName() in dynamic view rendering means:

If $billing is an Invoice, it calls Invoice::getViewName()
If $billing is a Bill, it calls Bill::getViewName()

This is classic polymorphism:

"Different objects, same method, different behavior."